requirements.txt
Flask==2.3.3
pytesseract==0.3.10
Pillow==9.5.0

